change README

+++ mandelbrot.py
@@ -8,55 +8,46 @@ LINUX = True if 'linux' in sys.platform else False
 WINDOWS = True if 'win' in sys.platform else False
 if (not (LINUX ^ WINDOWS)):
     raise "{} platform is not supported".format(sys.platform)
+workDir = os.getcwd()
 
-def getpng(n,c,name, compiled=False):
-    workDir = os.getcwd()
+def getpng(n,c,name):
     if os.path.exists(os.path.join(workDir, name)):
         # 已生成图像
+        img = cv2.imread(name)
+        cv2.imwrite(name,img)
+        return img
+    else:
+        # 未生成图像
+        if os.path.exists(os.path.join(workDir, 'Mandelbrot_python.out')) or os.path.exists(os.path.join(workDir, "Mandelbrot_python.exe")):
+            # 已编译且存在可执行文件
+            if WINDOWS:
+                cmd = 'Mandelbrot_python {} {} {} {}'
+            elif LINUX:
+                cmd = './Mandelbrot_python.out {} {} {} "{}"'
+            print(cmd.format(n, c.real, c.imag, name))
+            print('Running...')
+            os.system(cmd.format(n, c.real, c.imag, name))
+            if not os.path.exists(os.path.join(workDir, name)):
+                raise "Failed to generate the image\n生成图片失败"
             img = cv2.imread(name)
             cv2.imwrite(name,img)
             return img
-    else:
-        # 未生成图像
-        if compiled:
-            # 可执行文件已编译
-            if os.path.exists(os.path.join(workDir, 'Mandelbrot_python.out')) or os.path.exists(os.path.join(workDir, "Mandelbrot_python.exe")):
-                # 已编译且存在可执行文件
-                if WINDOWS:
-                    # os.system('Mandelbrot_python '+str(n)+' '+str(c.real)+' '+str(c.imag)+' '+name)
-                    cmd = 'Mandelbrot_python {} {} {} {}'
-                if LINUX:
-                    # os.system('./Mandelbrot_python.out '+str(n)+' '+str(c.real)+' '+str(c.imag)+' '+name)
-                    cmd = './Mandelbrot_python.out {} {} {} "{}"'
-                print(cmd.format(n, c.real, c.imag, name))
-                print('Running...')
-                os.system(cmd.format(n, c.real, c.imag, name))
-                img = cv2.imread(name)
-                cv2.imwrite(name,img)
-                return img
-            else:
-                raise "编译失败"
         else:
             # 可执行文件未编译
             print('No executable file!\nCompiling...')
-            if LINUX:
-                cmd = 'gcc Mandelbrot_python.c -o Mandelbrot_python.out -lm'
-            elif WINDOWS:
+            if WINDOWS:
                 cmd = 'gcc -lm Mandelbrot_python.c -o Mandelbrot_python.exe'
+            elif LINUX:
+                cmd = 'gcc Mandelbrot_python.c -o Mandelbrot_python.out -lm'
             print(cmd)
             os.system(cmd)
+            if not (os.path.exists(os.path.join(workDir, 'Mandelbrot_python.out')) or os.path.exists(os.path.join(workDir, "Mandelbrot_python.exe"))):
+                raise "Failed to compile\n编译失败"
             return getpng(n ,c, name, True)
 
 
 
 if __name__=='__main__':
-    from datetime import datetime
-
-    # try:
-    #     os.mkdir('runs/')
-    # except:
-    #     pass
-    # dirname='runs/'+datetime.now().strftime('%Y-%m-%d_%Hh%Mm%Ss')+'/'
     dirname = 'images/'
     try:
         os.mkdir(dirname)
@@ -72,12 +63,12 @@ if __name__=='__main__':
             n+=1
             print(str(n)+','+str(c+((x-127.5)+(127.5-y)*1j)/64/2**n))
 
-    img=getpng(n,c,dirname+str(n)+'_'+str(c)+'.png')
+    img = getpng(n,c,dirname+str(n)+'_'+str(c)+'.png')
     cv2.namedWindow('image')
     cv2.setMouseCallback('image',enlarge)
 
     while(True):
-        img=getpng(n,c,dirname+str(n)+'_'+str(c)+'.png', True)
+        img=getpng(n,c,dirname+str(n)+'_'+str(c)+'.png')
         cv2.imshow('image',img)
         if cv2.waitKey(100)&0xFF==27:
             break
